# 接口文档/openapi.yaml
openapi: 3.1.0
info:
  title: 信创智能客服公共 API
  version: 1.0.0
  description: 用户 Web 端公共接口。管理端和内部接口的人读契约见《智能客服系统接口文档.md》。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PasswordLogin'
                - $ref: '#/components/schemas/SmsLogin'
              discriminator: { propertyName: mode }
      responses:
        '200':
          description: 登录成功，同时设置 XC_REFRESH_TOKEN HttpOnly Cookie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSessionResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security:
        - refreshCookie: []
      responses:
        '200':
          description: refresh token 轮换成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSessionResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security:
        - refreshCookie: []
      responses:
        '204': { description: 已注销当前设备 }
        default: { $ref: '#/components/responses/Problem' }
...
  /chat/stream:
    post:
      tags: [Chat]
      operationId: streamChat
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatStreamRequest' }
      responses:
        '200':
          description: SSE 事件流，事件结构见 sse-events.schema.json
          headers:
            Cache-Control:
              schema: { type: string, const: no-cache }
            X-Accel-Buffering:
              schema: { type: string, const: 'no' }
          content:
            text/event-stream:
              schema: { type: string }
        default: { $ref: '#/components/responses/Problem' }
...
  /chat/requests/{requestId}/cancel:
    post:
      tags: [Chat]
      operationId: cancelChatRequest
      parameters:
        - $ref: '#/components/parameters/RequestIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 当前请求状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeBase'
                  - properties: { data: { $ref: '#/components/schemas/ChatRequestResult' } }
        default: { $ref: '#/components/responses/Problem' }
...
  /tickets/{ticketId}:
    get:
      tags: [Tickets]
      operationId: getTicket
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 本人工单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeBase'
                  - properties: { data: { $ref: '#/components/schemas/Ticket' } }
        default: { $ref: '#/components/responses/Problem' }
  /tickets/{ticketId}/replies:
    post:
      tags: [Tickets]
      operationId: replyTicket
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 1000 }
                attachmentIds:
                  type: array
                  maxItems: 5
                  items: { type: string }
      responses:
        '201':
          description: 回复已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeBase'
                  - properties: { data: { $ref: '#/components/schemas/TicketReply' } }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: XC_REFRESH_TOKEN
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    RequestIdPath:
      name: requestId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    TicketId:
      name: ticketId
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: RFC 7807 业务错误
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    EnvelopeBase:
      type: object
      required: [requestId, data]
      properties:
        requestId: { type: string }
        data: {}
    Problem:
      type: object
      required: [title, status, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PasswordLogin:
      type: object
      required: [mode, account, password, rememberDevice]
      properties:
        mode: { type: string, const: password }
        account: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, minLength: 6, maxLength: 128, format: password }
        rememberDevice: { type: boolean }
    SmsLogin:
      type: object
      required: [mode, phone, code, rememberDevice]
      properties:
        mode: { type: string, const: sms }
        phone: { type: string, pattern: '^1[0-9]{10}$' }
        code: { type: string, pattern: '^[0-9]{6}$' }
        rememberDevice: { type: boolean }
    AuthUser:
      type: object
      required: [id, name, phone, avatarText]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string, description: 脱敏手机号 }
        avatarText: { type: string, maxLength: 2 }
    AuthSessionResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - properties:
            data:
              type: object
              required: [accessToken, expiresAt, user]
              properties:
                accessToken: { type: string }
                expiresAt: { type: string, format: date-time }
                user: { $ref: '#/components/schemas/AuthUser' }
...
    Citation:
      type: object
      required: [title, snippet, sourceLocator]
      properties:
        title: { type: string }
        sourceId: { type: integer, format: int64 }
        snippet: { type: string }
        sourceLocator: { type: string }
        page: { type: integer, minimum: 1 }
...
    ChatStreamRequest:
      type: object
      required: [message]
      properties:
        sessionId: { type: integer, format: int64, minimum: 1, description: 首次会话必须省略 }
        message: { type: string, minLength: 1, maxLength: 8000 }
    ChatRequestResult:
      type: object
      required: [status, sessionId, citations, error]
      properties:
        status: { type: string, enum: [ACCEPTED, RUNNING, SUCCEEDED, FAILED, CANCELLED, INTERRUPTED] }
        sessionId: { type: integer, format: int64 }
        assistantMessageId: { type: integer, format: int64 }
        answer: { type: string, description: 仅 SUCCEEDED 返回 }
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        error:
          type: [object, 'null']
          properties:
            code: { type: string }
            message: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
...
    TicketStatus:
      type: string
      enum: [PENDING, PROCESSING, WAITING_USER, RESOLVED, CLOSED]
    TicketAttachment:
      type: object
      required: [id, fileName, size, contentType]
      properties:
        id: { type: string }
        fileName: { type: string }
        size: { type: integer, minimum: 0 }
        contentType: { type: string }
    TicketReply:
      type: object
      required: [id, sender, senderName, content, createdAt]
      properties:
        id: { type: string }
        sender: { type: string, enum: [user, admin] }
        senderName: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    TicketTimelineItem:
      type: object
      required: [id, title, description, status, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        createdAt: { type: string, format: date-time }
    Ticket:
      type: object
      required: [id, title, category, deviceBrand, deviceModel, description, contact, status, createdAt, updatedAt, attachments, replies, timeline]
      properties:
        id: { type: string }
        title: { type: string }
        category: { type: string }
        deviceBrand: { type: string }
        deviceModel: { type: string }
        description: { type: string }
        contact: { type: string, description: 返回时必须脱敏 }
        status: { $ref: '#/components/schemas/TicketStatus' }
        assignee: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/TicketAttachment' }
        replies:
          type: array
          items: { $ref: '#/components/schemas/TicketReply' }
        timeline:
          type: array
          items: { $ref: '#/components/schemas/TicketTimelineItem' }
...